The assumption made about image paths for WordPress.com blogs is incorrect. Files aren't necessarily uploaded with the exact same filenames, so as written, this can (and likely will) result in broken image urls.
This change causes the image urls to be rewritten only when the $Type is .org.

WordPress.com will now auto-import images from the perlbal.hi-pi.com domain.

It might be beneficial to export the attachment xml only for posts and pages,
since now we will import the images specified in the xml file and also
auto-import the images referenced within the post content.

This PR also fixes at least some of the encoding issues for accented characters.
I tested with an import that produced malformed characters previously, and
an export file generated with these changes has titles with the correct accents.